Add pytest coverage for dns_file.py and refactor record handling
- Implement tests for DNSFile class to ensure correct parsing and handling of DNS zone files.
- Add fixtures for sample DNS records including A, CNAME, and PTR records.
- Refactor DNSFile to improve record management and error handling.
- Remove deprecated records.py file and update imports accordingly.
