+++ tests/test_record_types.py
@@ -1,13 +1,13 @@
 from cleandns.record_types import PTRRecord, RecordType, DNSClass
 
-def test_a_record_creation(sample_a_record):
-    """Test the creation of an A record."""
-    assert sample_a_record.name == "www.example.com"
-    assert sample_a_record.ttl == 3600
-    assert sample_a_record.type == RecordType.A
-    assert sample_a_record.class_ == DNSClass.IN
-    assert sample_a_record.rdata == "1.2.3.4"
-    assert sample_a_record.comment == "Sample A record"
+#def test_a_record_creation(sample_a_record):
+#    """Test the creation of an A record."""
+#    assert sample_a_record.name == "www.example.com"
+#    assert sample_a_record.ttl == 3600
+#    assert sample_a_record.type == RecordType.A
+#    assert sample_a_record.class_ == DNSClass.IN
+#    assert sample_a_record.rdata == "1.2.3.4"
+#    assert sample_a_record.comment == "Sample A record"
 
 #def test_ptr_record_sorting():
 #    """Test that PTR records sort numerically by IP segments, not alphabetically."""
